bot.py
import os
from webdav3.client import Client
import discord
import asyncio

# Environment Variables
WEBDAV_URL = os.environ.get('WEBDAV_URL')
WEBDAV_LOGIN = os.environ.get('WEBDAV_LOGIN')
WEBDAV_PASSWORD = os.environ.get('WEBDAV_PASSWORD')
DISCORD_TOKEN = os.environ.get('DISCORD_TOKEN')

# Connect to WebDAV
webdav_client = Client(webdav_options)

# Discord Bot Configuration
TOKEN = 'your_discord_bot_token'
client = discord.Client()

# Function to check for new files
async def check_for_new_files():
    while True:
        # List files in WebDAV directory
        files = webdav_client.list('/')
        # TODO: Implement logic to determine new or updated files
        
        # Placeholder for posting to Discord
        # TODO: Post notification to Discord channel

        # Wait for 5 minutes before next check
        await asyncio.sleep(300)

@client.event
async def on_ready():
    print(f'Logged in as {client.user}')
    client.loop.create_task(check_for_new_files())

client.run(TOKEN)
