requirements.txt
discord.py
webdavclient3
aiohttp